Refactor to short timestamps
Close #43

+++ app/models/card.rb
@@ -1,6 +1,6 @@
 class Card
   include Mongoid::Document
-  include Mongoid::Timestamps
+  include Mongoid::Timestamps::Short
   include Domainable
   include Sluggable
